requirements.txt
paramiko
polars
packaging
pydantic
click
hvac
pyyaml
psycopg[binary]
sqlalchemy
